further simplifying VictionProcessor
