+++ core/state_processor.go
@@ -71,7 +71,7 @@ func NewStateProcessor(config *params.ChainConfig, bc *BlockChain, engine consen
 // transactions failed to execute due to insufficient gas it will return an error.
 func (p *StateProcessor) Process(block *types.Block, statedb *state.StateDB, cfg vm.Config) (types.Receipts, []*types.Log, uint64, error) {
 	// Viction hooks
-	if err := p.viction.BeforeBlockProcess(block, statedb); err != nil {
+	if err := p.viction.PreBlockProcess(block, statedb); err != nil {
 		return nil, nil, 0, err
 	}
 	var (
@@ -93,13 +93,13 @@ func (p *StateProcessor) Process(block *types.Block, statedb *state.StateDB, cfg
 		if err != nil {
 			return nil, nil, 0, err
 		}
-		if err := p.viction.BeforeApplyTransaction(block, tx, msg, statedb); err != nil {
+		if err := p.viction.PreApplyTransaction(block, tx, msg, statedb); err != nil {
 			return nil, nil, 0, err
 		}
 		statedb.Prepare(tx.Hash(), block.Hash(), i)
 
 		// Apply Viction-specific system transactions (BlockSigner, native trading/lending).
-		handled, receipt, _, err, _ := p.viction.ApplyNativeTransaction(statedb, tx, header, usedGas)
+		handled, receipt, _, err, _ := p.viction.ApplyNativeTransaction(tx, header, statedb, usedGas)
 		if err != nil {
 			return nil, nil, 0, err
 		}
@@ -112,7 +112,8 @@ func (p *StateProcessor) Process(block *types.Block, statedb *state.StateDB, cfg
 		}
 
 		// Execute Viction-specific post-transaction logic.
-		if err := p.viction.AfterApplyTransaction(tx, msg, statedb, receipt, receipt.GasUsed, err); err != nil {
+		failed := receipt.Status == types.ReceiptStatusFailed
+		if err := p.viction.PostApplyTransaction(tx, msg, statedb, receipt.GasUsed, failed); err != nil {
 			return nil, nil, 0, err
 		}
 		receipts = append(receipts, receipt)
@@ -122,7 +123,7 @@ func (p *StateProcessor) Process(block *types.Block, statedb *state.StateDB, cfg
 	p.engine.Finalize(p.bc, header, statedb, block.Transactions(), block.Uncles())
 
 	// Execute Viction-specific post-processing logic.
-	if err := p.viction.AfterProcess(block, statedb); err != nil {
+	if err := p.viction.PostBlockProcess(block, statedb); err != nil {
 		return nil, nil, 0, err
 	}
 
@@ -192,7 +193,7 @@ func applyTransaction(msg types.Message, config *params.ChainConfig, bc ChainCon
 // chain_makers) path.
 func ApplyTransaction(config *params.ChainConfig, bc ChainContext, author *common.Address, gp *GasPool, statedb *state.StateDB, header *types.Header, tx *types.Transaction, usedGas *uint64, cfg vm.Config) (*types.Receipt, error) {
 	// POSV: Viction system transactions bypass — same logic as block import.
-	if handled, receipt, _, err, _ := (&VictionProcessor{config: config}).ApplyNativeTransaction(statedb, tx, header, usedGas); handled {
+	if handled, receipt, _, err, _ := (&VictionProcessor{config: config}).ApplyNativeTransaction(tx, header, statedb, usedGas); handled {
 		return receipt, err
 	}
