+++ eth/api_tracer.go
@@ -504,7 +504,7 @@ func (api *PrivateDebugAPI) traceBlock(ctx context.Context, block *types.Block, 
 		msg, _ := tx.AsMessage(signer)
 		// Reproduce Viction per-tx pre-checks (balance override + blacklist) on the
 		// shared state before snapshotting it for the tracer, mirroring block import.
-		if err := vp.BeforeApplyTransaction(block, tx, msg, statedb); err != nil {
+		if err := vp.PreApplyTransaction(block, tx, msg, statedb); err != nil {
 			failed = err
 			break
 		}
@@ -522,7 +522,10 @@ func (api *PrivateDebugAPI) traceBlock(ctx context.Context, block *types.Block, 
 		}
 		// Decrement the running fee pool exactly as block import does, so the next
 		// task's copy starts from post-drain capacities.
-		vp.ProcessZeroGas(statedb, tx, msg.From(), res.UsedGas, res.Failed())
+		if err := vp.PostApplyTransaction(tx, msg, statedb, res.UsedGas, res.Failed()); err != nil {
+			failed = err
+			break
+		}
 		// Finalize the state so any modifications are written to the trie
 		// Only delete empty objects if EIP158/161 (a.k.a Spurious Dragon) is in effect
 		statedb.Finalise(vmenv.ChainConfig().IsEIP158(block.Number()))
@@ -636,7 +639,7 @@ func (api *PrivateDebugAPI) standardTraceBlockToFile(ctx context.Context, block 
 		}
 		// Reproduce Viction per-tx pre-checks (balance override + blacklist) before
 		// execution, mirroring block import.
-		if err = vp.BeforeApplyTransaction(block, tx, msg, statedb); err != nil {
+		if err = vp.PreApplyTransaction(block, tx, msg, statedb); err != nil {
 			return dumps, err
 		}
 		// Execute the transaction and flush any traces to disk
@@ -651,7 +654,9 @@ func (api *PrivateDebugAPI) standardTraceBlockToFile(ctx context.Context, block 
 		}
 		if err == nil {
 			// Decrement the running fee pool exactly as block import does.
-			vp.ProcessZeroGas(statedb, tx, msg.From(), res.UsedGas, res.Failed())
+			if err := vp.PostApplyTransaction(tx, msg, statedb, res.UsedGas, res.Failed()); err != nil {
+				return dumps, err
+			}
 		}
 		if err != nil {
 			return dumps, err
@@ -909,7 +914,7 @@ func (api *PrivateDebugAPI) computeTxEnv(block *types.Block, txIndex int, reexec
 		msg, _ := tx.AsMessage(signer)
 		// Reproduce Viction per-tx pre-checks (balance override + blacklist) before
 		// executing or returning, mirroring block import.
-		if err := vp.BeforeApplyTransaction(block, tx, msg, statedb); err != nil {
+		if err := vp.PreApplyTransaction(block, tx, msg, statedb); err != nil {
 			return nil, vm.BlockContext{}, nil, nil, err
 		}
 		txContext := core.NewEVMTxContext(msg)
@@ -924,7 +929,9 @@ func (api *PrivateDebugAPI) computeTxEnv(block *types.Block, txIndex int, reexec
 			return nil, vm.BlockContext{}, nil, nil, fmt.Errorf("transaction %#x failed: %v", tx.Hash(), err)
 		}
 		// Decrement the running fee pool exactly as block import does.
-		vp.ProcessZeroGas(statedb, tx, msg.From(), res.UsedGas, res.Failed())
+		if err := vp.PostApplyTransaction(tx, msg, statedb, res.UsedGas, res.Failed()); err != nil {
+			return nil, vm.BlockContext{}, nil, nil, err
+		}
 		// Ensure any modifications are committed to the state
 		// Only delete empty objects if EIP158/161 (a.k.a Spurious Dragon) is in effect
 		statedb.Finalise(vmenv.ChainConfig().IsEIP158(block.Number()))
